Allow Streams to be instantiated using GD resources
I've been working on a patch to laminas-barcode to create a Stream from
a GD resource, but am unable to, as the resource type has to be a
stream. However, after talking to mwop, some time ago, he suggested that
it'd be fine to patch Diactoros to allow for GD resource types to
instantiate streams as well; hence, this patch.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>