Refactor StreamTrait functionality into Stream.php
Following on from the advice of @lcobucci, this change refactors the
StreamTrait functionality into Stream, and removes the trait from both
Stream and StreamFactory. On thinking about it, it makes most sense to
have this functionality directly in Stream, and not, in effect,
duplicated in multiple locations by another form.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>