fix: allow strict checks for Content-* headers
Previously, we would check any key starting with `CONTENT_` in the `$_SERVER` array, and treat it as a header.
However, per #63, this is incorrect, and we should only look for `CONTENT_TYPE`, `CONTENT_LENGTH`, and `CONTENT_MD5`.
To keep backwards compatibility, this patch implements a switch, using the ENV variable LAMINAS_DIACTOROS_STRICT_CONTENT_HEADER_LOOKUP.
When this value is found in `$_SERVER` (which aggregates ENV variables as well), the logic for identifying headers in the `$_SERVER` array will become strict.

For version 3.0, we will make the strict lookup the default.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>