docs: link to org-wide documents
Instead of bringing the org-wide documents into the repository, the goal is to link to the org-wide documents from the `README.md` file of each repository.
By placing the content at the end of the file, we can automate adding the information, and it's in a location users will likely look for it.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>