security: add request filter functionality for handling x-forwarded-* headers
We have always accepted `X-Forwarded-*` headers as valid when generating a URI instance via `marshalUriFromSapi()` (and, by extension, `ServerRequestFactory::fromGlobals()`).
However, these headers should only be used if there is a trusted proxy in front of the application.

On the flip side, changing the behavior now would result in breakage of existing applications built on this library.

This patch does the following.

First, it creates a new function `Laminas\Diactoros\marshalUriFromSapiSafely()`.
This function ignores the `X-Forwarded-*` headers when generating the URI.
`ServerRequestFactory::fromGlobals()` now uses this new function internally, and `Laminas\Diactoros\marshalUriFromSapi()` has been marked deprecated.

Next, we have added `Laminas\Diactoros\RequestFilter\RequestFilterInterface`, which defines the single method `filterRequest(ServerRequestInterface $request): ServerRequestInterface`.
`ServerRequestFactory::fromGlobals()` now optionally accepts an implementation, and passes the request instance it generates to it before returning a request.
If no instance is provided, for version 2, it will create an instance of a new implementation, `LegacyXForwardedHeaderFilter` marked to trust any proxy; for version 3, it will create and use `NoOpRequestFilter`.

`LegacyXForwardedHeaderFilter` defines additional methods for the purpose of configuring what proxies to trust:

- `trustAny()` indicates all proxies (and all `X-Forwarded-*` headers) should be trusted.
- `trustProxies(string|string[] $proxyAddresses, string[] $trustedHeaders = LegacyXForwardedHeaderFilter::X_FORWARDED_HEADERS)` allows indicating one or more proxies that are trusted, as well as which `X-Forwarded-*` headers to trust.
  Proxy addresses may be specified as absolute IP addresses, or as CIDR subnet masks (e.g., 192.168.0.0/16, 10.0.0.0/24, etc.).

When invoked, if the proxy is trusted, the filter loops through the trusted headers and sees if any are present in the request; if so, it will update the URI with matched values, and return a new request instance with the new URI instance.

This approach allows us to expand functionality to address [RFC 7239](https://datatracker.ietf.org/doc/html/rfc7239) at a later date.

**The decision to make this opt-in versus secure-by-default for version 2 is to prevent a BC break for end-users who relied on the fact that we used these `X-Forwarded-*` headers when constructing the URI.**
For version 3, we will have secure-by-default settings.

We will need to add functionality to the [mezzio/mezzio](https://github.com/mezzio/mezzio) package to support using a `RequestFilterInterface` instance with the server request factory within the current v2 series, and could pontentially update the mezzio skeleton immediately to use the `NoOpRequestFilter` by default.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>