refactor: remove obsolete type checks where possible
Due to updating to use PSR-7 v2, we can remove a number of conditionals that did type-checking, and also update constructors to use union types to fully type the API exposed.
Doing so removes the need for some unit tests, as type errors will now be guaranteed by the runtime.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>