qa: updated Psalm baseline
Fixed one type issue in a test where it was possible to expand types expected.
All other reported issues are now due to type inference being less strict than the engine, leading us to perform validation to be correct, but Psalm complaining that the expected types would not require validation.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>