refactor: only allow integer, not float, keys
String float values are not cast to floats when used as array keys, unlike integers, so we do not need to treat them differently.
What that means, though, is that we _should_ manage detection of list arrays versus associative arrays, but otherwise allow integer array keys.
This patch does the following:

- Adds detection of lists passed for the headers array into `MessageTrait::setHeaders()`.
  - This is accomplished with a polyfill for `array_is_list`, that also considers an empty array as valid.
- Removes the `is_numeric()` check for the header name, replacing it with `is_int()`.
- Only cast the header name to a string before appying the regex if it is an int.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>